To split the methods, fixing box score issure from the refactor
